chores(quick-deploy): add quick deploy
This commits enable a quick deploy mechanis for SW360-Portlet. This mechanism consists of the following parts:
(1) Webapp configuration to prevent tomcat from caching JSPs
(2) Using liferay developer properties for development
(3) Mounting the development sources into the sw360 container
(4) Running a watch script which updates the sources in the webapp

Details:

(1) This feature is implemented in the core SW360 repository by deploying a modified context.xml when developing.

(2) Liferay provides a property file named "portal-developer.properties" which must be explicitly activated. This
property file contains useful properties for development such as deactivating the cache for static resources. In
production mode static resources are cached per session.
If you create the containers without the flag "--prod", these developer properties will now be loaded. As a result
static resources can be updated in the webapps folder and will no longer be cached.

(3) In order to update the webapps folder of the portlet in tomcat it is necessary to access the developer workspace.
Therefore the folder "sw360-portlet/src/webapp" folder is mapped to folder "/opt/sw360/quickdeploy" by default. The
default setting assumes that you have checked out the repositories using the following structure:

    workspacedir
        |
        |- sw360
        |- sw360-chores

If your core SW360 repository is located somewhere else you might set the environment variable "SW360_REPO_PATH" to
your checked out repository.

(4) The folder "/opt/sw360/quickdeploy" is a mirror of the development workspace. The changed sources must be copied
to the webapps folder. For this purpose the script "quickdeploy.sh" is deployed to the container and started by default.
This scripts watches "/opt/sw360/quickdeploy" and copies every changed source to the appropiate webapps folder of the
portlet. As a result a simple refresh in the browser is enough to load the static resources or JSPs.
In order to disable QuickDeploy you might start the "sw360chores.pl" script with the switch --no-quick-deploy.

Signed-off-by: Leo von Klenze <leo.vonklenze@scansation.de>